Feature attendance functionality

+++ public/javascript/comment.js
@@ -2,8 +2,8 @@ import getUrl from "./config.js"
 var comment = document.getElementById('commentform');
 comment.addEventListener('submit', submitCommentForm);
 
-const comment_rm_btn = document.getElementById('comment-rm-btn');
-comment_rm_btn.addEventListener('click', removeCommentbyid);
+const commentsEl = document.getElementById("all-comments-1248");
+commentsEl.addEventListener("click", commentSetting);
 
 function submitCommentForm(event) {
 
@@ -34,18 +34,17 @@ function submitCommentForm(event) {
     }).then(() => { location.reload() })
 }
 
-function removeCommentbyid(event) {
+function commentSetting(event) {
 
-    event.preventDefault()
-    console.log("hello comment delete")
-    let comment_id = comment_rm_btn.dataset.commentid;
-    let creator_id = comment_rm_btn.dataset.creatorid;
-    console.log("creator_id ", creator_id);
-    console.log("comment_id ", comment_id);
-    var data = {
-        creator_id: creator_id,
-    }
-    fetch(`http://localhost:3000/api/comments/${comment_id}`, {
+    var optionSelected = event.target;
+    
+    if (optionSelected.matches(".comment-rm-btn-1248")) {
+
+      let data = {
+        creator_id: optionSelected.dataset.creatorid,        
+      };
+      let comment_id = optionSelected.dataset.commentid;
+       fetch(`http://localhost:3000/api/comments/${comment_id}`, {
         credentials: "include",
         method: "DELETE",
         headers: {
@@ -53,5 +52,8 @@ function removeCommentbyid(event) {
             "Accept": "application/json"
         },
         body: JSON.stringify(data)
-    }).then(() => { })
+
+    }).then(() => { location.reload() })
+    };
+
 }
